+++ tests/prompt/tree/prompt-tree-edge_test.py
@@ -6,6 +6,7 @@ Unit Tests (using pytest) for:
 load_corpus_tree
 """
 
+from pathlib import Path
 from unittest.mock import mock_open, patch
 
 from kaye_engine.prompt.prompt_corpus_loader import load_corpus_tree
@@ -19,7 +20,7 @@ class TestEdge:  # various edge cases
         m = mock_open(read_data="")
 
         with patch("builtins.open", m):
-            tree = load_corpus_tree("edge-empty1", "dummy-path.md")
+            tree = load_corpus_tree("edge-empty1", [Path("dummy-path.md")])
             assert tree.depth == 0
             assert tree.parent is None
 
@@ -27,7 +28,7 @@ class TestEdge:  # various edge cases
         m = mock_open(read_data="\n")
 
         with patch("builtins.open", m):
-            tree = load_corpus_tree("edge-empty2", "dummy-path.md")
+            tree = load_corpus_tree("edge-empty2", [Path("dummy-path.md")])
             assert tree.depth == 0
             assert tree.parent is None
 
@@ -35,6 +36,6 @@ class TestEdge:  # various edge cases
         m = mock_open(read_data="\n" * 10)
 
         with patch("builtins.open", m):
-            tree = load_corpus_tree("edge-empty3", "dummy-path.md")
+            tree = load_corpus_tree("edge-empty3", [Path("dummy-path.md")])
             assert tree.depth == 0
             assert tree.parent is None