+++ tests/prompt/tree/prompt-tree-empty_test.py
@@ -6,6 +6,7 @@ Unit Tests (using pytest) for:
 load_corpus_tree
 """
 
+from pathlib import Path
 from unittest.mock import mock_open, patch
 
 import pytest
@@ -44,7 +45,9 @@ More text.
     )
 
     with patch("builtins.open", m):
-        return load_corpus_tree("prompt-tree-fenced-test", "dummy-path.md")
+        return load_corpus_tree(
+            "prompt-tree-fenced-test", [Path("dummy-path.md")]
+        )
 
 
 @pytest.fixture(scope="session")
@@ -100,7 +103,9 @@ This project is licensed under the MIT License.
 """)
 
     with patch("builtins.open", m):
-        return load_corpus_tree("prompt-tree-empty-test", "dummy-path.md")
+        return load_corpus_tree(
+            "prompt-tree-empty-test", [Path("dummy-path.md")]
+        )
 
 
 # pytest #######################################################################